Added new dontShowLogDialog method to the ChangeLog class.
This feature is useful when you want to not show a change log dialog for the current app version.
A good reason to do this is if you are showing a new user experience on initial install of the app
and do not want to show the change log except for on subsequent app updates.