Added method getLogArray to obtain an array of ReleaseItem.
Added a new public method, getLogArray(full), to obtain a SparseArray of ReleaseItem objects, having the versioncode as the key, and the ReleaseItem object as value, with the versionname, versioncode and list of changes.
The new method uses the same code than getLog(full) to parse the chagelogs and merge localized and default xmls. Now getLog reuses this new method to avoid code duplicates.

Modified class ReleaseItem to be public, and its properties encapsulated to be read-only.